#151 - refactoring: added package
