+++ src/main/java/org/opengis/cite/ogcapifeatures10/conformance/crs/discovery/DiscoveryCollection.java
@@ -1,9 +1,6 @@
-package org.opengis.cite.ogcapifeatures10.conformance.crs;
+package org.opengis.cite.ogcapifeatures10.conformance.crs.discovery;
 
-import static org.opengis.cite.ogcapifeatures10.EtsAssert.assertDefaultCrs;
-import static org.opengis.cite.ogcapifeatures10.OgcApiFeatures10.DEFAULT_CRS;
-import static org.opengis.cite.ogcapifeatures10.OgcApiFeatures10.DEFAULT_CRS_WITH_HEIGHT;
-import static org.opengis.cite.ogcapifeatures10.util.JsonUtils.hasAtLeastOneSpatialFeatureCollection;
+import static org.opengis.cite.ogcapifeatures10.EtsAssert.assertValidCrsIdentifier;
 
 import java.util.List;
 
@@ -30,7 +27,7 @@ import io.restassured.path.json.JsonPath;
  *
  * @author <a href="mailto:goltz@lat-lon.de">Lyn Goltz </a>
  */
-public class DiscoveryCollectionDefaultCrs extends AbstractDiscoveryCollection {
+public class DiscoveryCollection extends AbstractDiscoveryCollection {
 
     /**
      * Test: crs property in the collection objects in the path /collections
@@ -41,12 +38,28 @@ public class DiscoveryCollectionDefaultCrs extends AbstractDiscoveryCollection {
     @Test(description = "Implements A.1 Discovery, Abstract Test 1 (Requirement /req/crs/crs-uri, /req/crs/fc-md-crs-list A, /req/crs/fc-md-storageCrs, /req/crs/fc-md-crs-list-global), "
                         + "crs property in the collection object in the path /collection", dataProvider = "collectionIdAndJson", dependsOnGroups = "crs-conformance")
     public void verifyCollectionCrsIdentifierOfCrsProperty( String collectionId, JsonPath collection ) {
-        Object extent = collection.get( "extent" );
-        if ( hasAtLeastOneSpatialFeatureCollection( extent ) ) {
-            List<String> crs = JsonUtils.parseAsList( "crs", collection );
-            assertDefaultCrs( crs,
-                              String.format( "Collection with id '%s' does not specify one of the default CRS '%s' or '%s' but provides at least one spatial feature collections",
-                                             collectionId, DEFAULT_CRS, DEFAULT_CRS_WITH_HEIGHT ) );
+        List<String> crs = JsonUtils.parseAsList( "crs", collection );
+        for ( String crsValue : crs ) {
+            assertValidCrsIdentifier( crsValue,
+                                      String.format( "Collection with id '%s' contains invalid CRS identifier property 'crs': '%s'",
+                                                     collectionId, crsValue ) );
+        }
+    }
+
+    /**
+     * Test: storageCrs property in the collection objects in the path /collections
+     *
+     * @param collection
+     *            the /collection object, never <code>null</code>
+     */
+    @Test(description = "Implements A.1 Discovery, Abstract Test 1 (Requirement /req/crs/crs-uri, /req/crs/fc-md-crs-list A, /req/crs/fc-md-storageCrs, /req/crs/fc-md-crs-list-global), "
+                        + "storageCrs property in the collection object in the path /collection", dataProvider = "collectionIdAndJson", dependsOnGroups = "crs-conformance")
+    public void verifyCollectionCrsIdentifierOfStorageCrsProperty( String collectionId, JsonPath collection ) {
+        List<String> crs = JsonUtils.parseAsList( "storageCrs", collection );
+        for ( String crsValue : crs ) {
+            assertValidCrsIdentifier( crsValue,
+                                      String.format( "Collection with id '%s' contains invalid CRS identifier property 'storageCrs': '%s'",
+                                                     collectionId, crsValue ) );
         }
     }
 }
