+++ src/main/java/org/opengis/cite/ogcapifeatures10/conformance/crs/discovery/DiscoveryCollections.java
@@ -1,4 +1,4 @@
-package org.opengis.cite.ogcapifeatures10.conformance.crs;
+package org.opengis.cite.ogcapifeatures10.conformance.crs.discovery;
 
 import static org.opengis.cite.ogcapifeatures10.EtsAssert.assertValidCrsIdentifier;
